Move copy_stack_trace to gof.opt
